start consolidating relation test files
