+++ libs/datamodel/core/tests/attributes/mod.rs
@@ -5,12 +5,7 @@ pub mod id_negative;
 pub mod id_positive;
 pub mod index;
 pub mod map;
-pub mod relations_basic;
-pub mod relations_consistency;
-pub mod relations_legacy;
-pub mod relations_negative;
-pub mod relations_new;
-pub mod relations_positive;
+pub mod relations;
 pub mod unique;
 pub mod unique_criteria;
 pub mod updated_at_negative;
