+++ libs/datamodel/core/tests/attributes/relations/mod.rs
@@ -1,0 +1,4 @@
+pub mod relations_consistency;
+pub mod relations_negative;
+pub mod relations_new;
+pub mod relations_positive;
