+++ libs/datamodel/core/tests/base/base_types.rs
@@ -189,3 +189,32 @@ fn json_type_must_work_for_some_connectors() {
         .assert_has_scalar_field("json")
         .assert_base_type(&ScalarType::Json);
 }
+
+#[test]
+fn resolve_enum_field() {
+    let dml = r#"
+    model User {
+        id Int @id
+        email String
+        role Role
+    }
+
+    enum Role {
+        ADMIN
+        USER
+        PRO
+    }
+    "#;
+
+    let schema = parse(dml);
+    let user_model = schema.assert_has_model("User");
+    user_model
+        .assert_has_scalar_field("email")
+        .assert_base_type(&ScalarType::String);
+    user_model.assert_has_scalar_field("role").assert_enum_type("Role");
+
+    let role_enum = schema.assert_has_enum("Role");
+    role_enum.assert_has_value("ADMIN");
+    role_enum.assert_has_value("PRO");
+    role_enum.assert_has_value("USER");
+}
